+++ Common/Header/Macro.h
@@ -10,7 +10,7 @@
 #define LK8000_MACRO_H
 
 // Be sure we are executing not faster than 1Hz!
-#define ONEHZLIMITER {;static double nextHB=0; if (LKHearthBeats< nextHB) return; nextHB=LKHearthBeats+2; }
+#define ONEHZLIMITER {;static unsigned nextHB=0; if (LKHearthBeats< nextHB) return; nextHB=LKHearthBeats+2; }
 
 // For debugging and trace tests, it should not be used out of testbench mode
 // and not committed to masterbranch normally, for individual usage during debugging
